Additional Scripts and Compression Improvements
compress.sh now logs time taken and splits blocks in parallel.

analyze.sh reports and logs the size of outputs.

workload.sh prepares and runs a YCSB workload on a redis database, then
obtains a core dump of the database.

+++ compress.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
-chunk_sizes=(256 512 1k 4k)
+if [ -z $1 ] && [ -z $2 ]; then
+  echo error: please enter core file name and proc limit
+  exit
+fi
+
+chunk_sizes=(4k 1k 512 256)
 
 rm -rf c$1
 mkdir c$1
@@ -8,21 +13,23 @@ mkdir c$1
 for cs in ${chunk_sizes[@]}; do
 
   # split core file into initial n blocks
-  mkdir c$1/core_${cs}
+  mkdir c$1/core_$cs
   split -n $2 $1 c$1/core_${cs}/_
-  cd c$1/core_${cs}
+  
+  mkdir c$1/core_${cs}/lz4
+  mkdir c$1/core_${cs}/gzip
+  touch ctimes
 
-  # split blocks into chunks
-  for f in _[a-z]*; do
-    mkdir _$f
-    split -b $cs $f _$f/$f
-    rm $f
-  done
-
-  mkdir lz4
-  mkdir gzip
-  cd ../..
 done
+
+split_all() {
+
+  # split each block into chunks
+  mkdir _$2
+  split -b $1 $2 _$2/$2
+  rm $2
+}
+export -f split_all
 
 lz4_all() {
 
@@ -40,7 +47,7 @@ gzip_all() {
   # gzip compress each chunk
   cd $1
   for f in *; do
-    gzip $f > ../gzip/$f.gz
+    gzip -c $f > ../gzip/$f.gz
   done
   cd ..
   rm -rf $1
@@ -49,19 +56,24 @@ export -f gzip_all
 
 for cs in ${chunk_sizes[@]}; do
 
-  cd c$1/core_${cs}
+  cd c$1/core_$cs
+
+  parallel split_all $cs ::: _[a-z]*
 
   startTime="$(date -u +%s)"
   parallel lz4_all ::: __*
   endTime="$(date -u +%s)"
   echo lz4 --- ${cs} --- $(($endTime-$startTime)) sec
+  echo lz4 --- ${cs} --- $(($endTime-$startTime)) sec >> ../ctimes
 
   startTime="$(date -u +%s)"
   parallel gzip_all ::: __*
   endTime="$(date -u +%s)"
   echo gzip --- ${cs} --- $(($endTime-$startTime)) sec
+  echo gzip --- ${cs} --- $(($endTime-$startTime)) sec >> ../ctimes
 
   cd ../..
 
 done
 
+mv $1 c$1/
